requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai-whisper==20231117
llama-cpp-python==0.2.77
python-multipart==0.0.9
httpx==0.27.0
